qdoc: Clear config variables before loading a new .qdocconf file
Previously, this problem was masked as the Config object was local
to the processQdocconf() function in main.cpp - but now we keep
that object and process multiple .qdocconf files in single-exec
mode.

When in single-exec mode, the previously read variables remained
in the config, resulting in all kinds of unintended consequences
- such as processing source directories from other documentation
projects.

Fixes: QTBUG-78776
Change-Id: Ic23d1cb0eacb112cc60a63cd43b4d51ff57f0583
Reviewed-by: Paul Wicking <paul.wicking@qt.io>