qdoc: Remove multiple QTypeInfo entries from All Classes page
QTypeInfo is a class that is generated by a Qt macro. It is not
documented and is not supposed to be documented, but clang sees
it as being in the public API. This update ensures that the
class is marked internal prior to generating the index file.
Ironically, this was one of the classes that motivated the
addition of the \dontdocument command, and it was the test for
isDontDocument() that caused this bug. We might have to visit
the \dontdocument command again.

Task-number: QTBUG-79088
Change-Id: I11307a2236e2ebcdcc205952056ddaca88fc0bd3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>