qdoc: Introduce ignorewords configuration variable
QDoc automatically generates hyperlinks for words that qualify for
auto-linking. This may become a problem whenever there is a qualified
word as a section or a page title, and that word appear in the
documentation frequently. For example, 'macOS' is mentioned often
and each occurrence is linked, which is unnecessary and sometimes
confusing.

Provide a way to exclude specified words from being auto-linked.
Create a new helper function, DocParser::isAutoLinkString(), to
determine whether a string qualifies, and replace duplicated
code with calls to that function.

Fixes: QTBUG-79135
Change-Id: Ie53fe7ca0692f7b8e10a5f2208df5cd2ee2aab1d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>