qdoc: Fix issues with shared comment nodes
QDoc stores <sharedcomment> elements to the .index correctly, but
the nodes that were sharing a comment were added after the
shared comment. This meant that a shared comment node (SCN) could
not find any of its members as they hadn't been read in yet.

Fix this by delaying the creation of a new SCN until we have a list
of sharing nodes.

Also fix other related issues:

 - Generate content also for non-function SCN members.

 - Make SCN adopt the genus of a member when adding one.

 - Fix Node::isPropertyGroup() to return true for actual property
   groups (not all QML properties that share a comment) and fix usage
   of that method.

 - Fix linking to property groups when searching by the group name.

 - Restore indentation of property group members both in the summary
   section and in the All Members file.

Fixes: QTBUG-79204
Change-Id: I1702f39b178f52444436b800d4fe9cb99f976a60
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>