Implement DocBook support for qdoc
The use of this code would be to overhaul the translation project of
the documentation in French, for Developpez.com (which is now stuck
at Qt 4.7 due to outdated tools: https://qt.developpez.com/doc/).
This means that this code will be used and maintained, whatever the
decision follows QTBUG-71493, for instance.

The DocBook support is not 100% vanilla DocBook, as some tags are
missing, but rather an extension (which could be officialised soon:
https://github.com/docbook/docbook/issues/111). These tags are used
to encode metadata about the code being commented (classes, fields,
functions, etc.). More precisely, the required tags are:
[enum|macro|namespace|typedef|union][|name|synopsis],
enum[value|identifier|item], [specialized]template[id].
These extensions can be enabled by the docbook-extensions parameter.

If you want to see the history:
https://github.com/dourouc05/qttools/tree/dourouc05-qdoc-docbook

Change-Id: I919976d8680b41c5ca69bdb79a0b824c29b89d05
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>