qdoc: Avoid excessive 'Illegal \reimp' warnings
In Generator::generateReimplementsClause(), QDoc tries to output
reimplementation information for function nodes, and intends to
print out a warning if there is no documented virtual function
in the base class to override.

This warning was triggered too often - especially, when the
base implementation was an \internal class, and the documented
reimplementation did not even use the \reimp command.

This commit fixes the warning to be omitted for \internal
base member functions.

Change-Id: Ic39327520b28e5034c489a4037297210fa8b10d9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>