qdoc: Ignore the 'enum' keyword when parsing a Q_PROPERTY type
QDoc's parsing of property types is slightly naive, in that it
cannot parse a type that consists of multiple words. These are
not common, but we do have a few instances of

Q_PROPERTY(enum EnumName ...)

in the Qt source. This commit allows QDoc to parse above properties
by ignoring the string 'enum'.

Fixes: QTBUG-80027
Change-Id: I3b01d463b081042dd27299c2121fafa3c99b98ce
Reviewed-by: Paul Wicking <paul.wicking@qt.io>