qdoc: Allow a documentation project to load all available index files
QDoc recognized '*' as a valid input for the 'depends' qdocconf
variable, with the intent to load all available index files. This did
not work as we cleared the resolved dependencies by re-fetching
the depends string from Config after resolving the dependencies.

This fix is a backport of another commit that is targeted for the
dev branch, but this useful already for 5.14.

Fixes: QTBUG-80051
Change-Id: I1d752208078452f4b16b1d855cdb67b8cbabf1f4
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>