qdoc: Ensure the output subdirectory is used for locating dependencies
Commit 2b6c90317 allowed QDoc to load all available index files if
'depends = *' configuration was used.

This could still fail when output from multiple modules is directed
to the same location (single-dir output mode). QDoc correctly adds
the output directory as a place to look for index files, but because
that output directory is also added as a dependency, it was removed
from the list of sub-dirs to look into once that dependency was
processed. This meant that all the subsequent dependencies could
not be located.

Also, there's a fundamental difference in how we must resolve
the dependencies; without .nosubdirs, QDoc takes the subdirectories
of -indexdir entries and treats them as dependencies. With
.nosubdirs, we need to look for actual .index files in the
-outputdir and load all of them.

Finally, move the logging line for 'depends = *' further down,
so that we can print out the number of resolved dependencies.

Task-number: QTBUG-80051
Change-Id: I35ff5c22e8e1f7e2b0c7de798531ad1c72813e51
Reviewed-by: Martin Smith <martin.smith@qt.io>