Qt Designer: Fix deprecated usage of function for DESIGNABLE
Use a property sheet instead, fixing warnings:
qdesigner_dockwidget_p.h:53: Warning: Providing a function for DESIGNABLE in a property declaration is deprecated and will not be supported in Qt 6 anymore.
qdesigner_dockwidget_p.h:54: Warning: Providing a function for DESIGNABLE in a property declaration is deprecated and will not be supported in Qt 6 anymore.

Change-Id: I9b3b09e5a6a13667782a8bdb7bdd3c4d7a6145c3
Reviewed-by: Lars Knoll <lars.knoll@qt.io>