Qt Designer: Handle properties that specify a function for DESIGNABLE
The behavior of Widgets will change to always return true for
those cases, which needs to be handled in the property sheet.

Add flags and types for handling the checked/checkable
dependencies to QDesignerPropertySheet::isEnabled().

QtWidgets also contained handling for the movable/allowedAreas
properties of QToolBar, but these appear to be non-functional, the
properties are designable regardless of parent. In addition, Qt
Designer currently only supports QToolBar on main windows.

Change-Id: Id417ee084a74ff91deaae13b567e64d8ff52c3e8
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>