qdoc: Restore alphabetical ordering of generated type lists
The comparison function used for sorting compares fullTitle() strings
for PageNode items. When the class hierarchy was refactored, this
had the unintended effect where this comparison was applied to lists
of C++ and QML types as well - the full title is empty for these,
so we ended up comparing the physical location of the type declarations
instead.

Fix this by preferring the fullName() over fullTitle() when comparing,
as the former is what the visible list item names are generated from.

Make the comparison function easier to read and maintain by introducing
a preprocessor macro.

Fixes: QTBUG-80259
Change-Id: I3f35aa80f1a90a6817f836146976d73b857df820
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>