qdoc: Fix regression in output for QML properties with shared comments
QDoc was incorrectly assuming that QML properties that share a comment
must belong to a property group.

Fix the issue, reduce duplication of code with functions that
output properties and methods, fix the 'id' attributes that were left
empty, and remove the <a name="..."> elements that are obsolete and
unnecessary now that 'id' has correct content.

Add a test for both properties and methods that share a comment, and
fix the expected output data for above changes.

Fixes: QTBUG-80768
Change-Id: Iddfca8214b6058987c76cd8f8dd6a842ce7cea89
Reviewed-by: Levon Sargsyan <levon.sargsyan@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>