qdoc: Introduce ignoresince configuration variable
Qt has a long history, with lots of classes and functions
documented being introduced in a specific version. This
\since information for decade-old features is not relevant
anymore, producing noise about ancient Qt releases that
have been long unsupported.

In preparation for Qt 6, introduce a feature where we can set
a cutoff value for \since information that can be ignored in
the generated documentation.

Fixes: QTBUG-81090
Change-Id: Ib35a2a4d8bc3a6cb9719223a45b02f0e63f337ea
Reviewed-by: Paul Wicking <paul.wicking@qt.io>