qdoc: Introduce locationinfo configuration variable
This simple boolean variable determines whether the full path
information of entities is stored into index files, and by extension,
the output of the WebXML generator.

The default value is true, and QDoc's output is unchanged. Using a
value of false is useful for testing QDoc; we can omit system
build paths in the output as those complicated the comparison of
expected and generated data.

[ChangeLog][qdoc] Introduced a configuration variable to drop
host-specific paths from the generated output.

Fixes: QTBUG-81091
Change-Id: Ic8df5e4a52a12821e2858a185c4e658905deb847
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>