[macdeployqt] Use entitlements from file for signing if present
Add the option to place a file with the suffix .entitlements into the
Contents/Resources subdirectory of an application bundle. The
entitlements listed in the file will be used for signing.

Bundles that are depoloyed as part of another bundle  (e.g. Helpers)
can specify a separate set of entitlements from the main bundle by
including an entitlements file of their own.

Only the first entitlement file per bundle will be used for signing as
the codesign tool will only use one entitlements file even when multple
are specified.

Task-number: QTBUG-77442
Change-Id: Iea356c1a70713f3a4b07281245a17fd7c87f6b11
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>