qdoc: Teach QDoc to output function and class template parameters
Add a visitor functor that reconstructs template parameters from
the Clang AST, and use it when either a class or a method template
is detected.

Store this information in the node - for now, they are used only
in the class reference page subtitle (for classes), and detailed
section of function documentation. Template parameters are not
considered when searching/linking to entities, they are for
display only.

[ChangeLog][qdoc] Added capability to display class/method template
parameters in the generated documentation.

Fixes: QTBUG-17456
Change-Id: I300cc63b9fa20d4f6efaeaa27ea3769635a7b32c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>