qdoc: Fix regression in QML types inheriting property groups
Property groups from base types were not listed on 'All Members'
pages of a QML type. This was because we skipped all shared
comment nodes for base types; property group is a special type
of SCN.

Remove duplication in the code that distributes the nodes; this
ensures that there is no difference in output between members of
a type and members inherited from its base types. Use logging
category to display debug warnings instead of qDebug().

Add a test to cover this.

Fixes: QTBUG-81262
Change-Id: I46c3e6afa87c510a7929088acbd1fb354f6733af
Reviewed-by: Paul Wicking <paul.wicking@qt.io>