qdoc: Handle \qmlmodule pages that are marked \internal
A special case of QML type documentation is a one where a type does
not require an import statement to be usable, and is not a basic type.
If the \inqmlmodule command was omitted, QDoc warned about it and
generated an incomplete import statement in the type reference.

We still want to see those warnings as they often are valid, but
allow the above usecase by marking a QML module \internal:

/*!
  \qmlmodule internal 1.0
  \internal
*/

/*!
  \qmltype SomeType
  \inqmlmodule internal
*/

Such a type will generate otherwise normal reference page, but without
an import statement. Also, the (internal) module name will not appear
in the generated file names (qml-sometype.html).

Fixes: QTBUG-81613
Change-Id: I50bb08f48fc4e7dbc7602ab20de39e039213810b
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>