qdoc: Ensure that \dontdocument works as expected
Class nodes marked with \dontdocument are not supposed to be seen
in the generated documentation. They already did not generate an
output page, but they still appeared as base classes in 'Inherits'
information as well as class listings, resulting in broken links.

Remove these nodes from all relevant lists. Handle Node::DontDocument
status in the .index files, and add a test to cover \dontdocument
usage.

Fixes: QTBUG-81468
Change-Id: I933817ff0ebf372cfa493a5ea824b682e0b1b160
Reviewed-by: Paul Wicking <paul.wicking@qt.io>