qdoc: Fix behavior with missing \inmodule command
QDoc warns about a missing \inmodule command and tells us that it has
selected a default (project) name for the module, but this had no
effect because we never added the node to the module.

Automatically adding the node to the module won't eliminate the
warnings, but ensures that a \module page whose name matches the
default one is populated with the classes/namespaces as expected.

Change-Id: Ib6a949cc2763c254d379ff99b08c60812f45e607
Reviewed-by: Paul Wicking <paul.wicking@qt.io>