qdoc: Refactor automatic note generation for function nodes
Function nodes may have a number of automatically generated notes
added to the documentation; as they share the same formatting,
collect them into a single virtual function in Generator, and
override it in DocBookGenerator.

Call the virtual function in Generator::generateBody() after
generating the text for the Doc instance. As we append the
notes to the body, the order in which the notes appear
may change slightly (if a function comment generates multiple
notes).

Add \note commands in the tests in preparation for a follow-up
commit that utilizes the refactored code.

Task-number: QTBUG-37355
Change-Id: Id92926e53b61e03a2d2feceda70987191b708975
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>