qdoc: Fix warning when accessing a QString outside of its range
Since skipSpacesOnLine() can end up making pos == len then we need to
check pos is less than length before querying input_ with that position.

This fixes the warning "Using QCharRef with an index pointing outside
the valid range of a QString. The corresponding behavior is deprecated,
and will be changed in a future version of Qt."

Fixes: QTBUG-80346
Change-Id: Ica1c57f616c3efaefd597ebb84e52b17325d22fa
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>