qdoc: Fix regression in .index file output
Commit 5234c7fc fixed an issue with missing \inmodule commands,
but it ended up adding non-aggregates into the module as well
which was not the intention. This caused excess nodes being
listed as a module members, causing issues for the WebXML output
format as well.

Partially revert 5234c7fc, fix the issue correctly, and update
the test for .index files.

Change-Id: I85fc19152b3de8993f92361fbe1ad2fd934c79ae
Reviewed-by: Paul Wicking <paul.wicking@qt.io>