Enforce zero warnings for QDoc's generatedoutput autotest
Fix all QDoc warnings for the tests and set the warning limit to zero.
If any warnings are introduced, QDoc will exit with non-zero value and
the test will fail.

The exceptions are the tests specific to the DocBook generator, where
we expect to see two warnings related to shared comment nodes. The
generator currently has trouble handling these.

Fixes: QTBUG-82808
Change-Id: I0482516e1e34b92592d921a4a0ffb41e4b2b4770
Reviewed-by: Paul Wicking <paul.wicking@qt.io>