qdoc: Fix regression in resolved example file paths
A recent refactoring of code related to generation of example file
lists caused a regression; if the parameter passed to an \example
command contained subdirectories, e.g.

  \example tutorials/gettingstarted

only the immediate example directory ('gettingstarted' above) was
recorded in generated example lists and manifest files.

Ensure that the file paths are prefixed with the full example
location, and add a test to cover this.

Fixes: QTBUG-83130
Change-Id: I061dcf6cd4e94a2c65e5a50a39f379759d7cd06f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>