lprodump: Fix crash when destructing moved-from ProFunctionDef
Consider the following code:

    auto x = ProFunctionDef(...);
    auto y = ProFunctionDef(...);
    x = std::move(y);

This crashed when destructing y, because the move c'tor nulls the
m_pro member and the d'tor calls m_pro->deref() without checking for
null.

The bug was introduced in qttools/4a700c90, but it is only triggered in cumulative
parsing mode, which lprodump uses, with the QHash rewrite.

Fixes: QTBUG-83765
Pick-to: 5.15
Change-Id: I85529547572ee3d9a2102fc0fa81e1c7b4b7f8a2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>