qdoc: Allow expansion of config variables
In addition to expanding environment variables within the values of
configuration variables, allow also the expansion from other config
variables. This gives us more flexibility in designing documentation
templates that can be customized per module.

The environment variables take priority - if expansion fails, do not
bail out immediately but mark the location for expansion once all
variables are read in.

For referencing variables, add support for a new syntax with enclosing
braces, with an optional parameter for controlling how multi-line
values are joined in the expanded string.

[ChangeLog][qdoc] Allow expansion from other configuration variables
in .qdocconf files.

Fixes: QTBUG-83826
Change-Id: Ifccd0c8844d65991df99cb8a2eb1d1bf2f3b4484
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>