qdoc: Refactor ConfigVar
There is little benefit in storing each encountered configuration variable
in a QMultiMap, together with the information about whether it appends or
overrides a previous variable.

Store the ConfigVar instances in a QMap instead, storing only the latest
state of each variable, as well as the path(s) where they're defined. We
need some logic for appending to existing variables, but traversing the
variable map is now much simpler.

Task-number: QTBUG-71176
Change-Id: Ic23111c6d3d8e2a29c5cec8ee2c8367e3146f444
Reviewed-by: Paul Wicking <paul.wicking@qt.io>