qdoc: Improve handling of templated types
The Clang visitor did not differentiate between templated class-like
types, always assuming that it is a 'class'. Fix this by retrieving
the type of the specialization the template refers to.

Ensure that templated structs and unions are named correctly in the
generated output.

Handle the type of the cursor that refers to a templated type alias,
and improve the code for resolving the template parameters for aliases.

[ChangeLog][qdoc] QDoc now correctly handles templated structs, unions,
and type aliases.

Fixes: QTBUG-67432
Change-Id: Ia0fbfe7b7f22a54270650d96d0d67a4b8b8fd182
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 4f97a639cc39e8b1a42181a662b1059cd9c1d581)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>