qdoc: Fix regression in HeaderNode name handling
Node::plainFullName() constructs a fully qualified name for a node,
traversing the chain of parents until it comes across the root
namespace or a HeaderNode - but if the start node was already a header,
it incorrectly returned an empty string as the node name.

Fix the issue and add a test for output generated by \headerfile command.

Pick-to: 5.15
Fixes: QTBUG-84070
Change-Id: Ib214c53be8ed0eea98708219a00ba43970cd4177
Reviewed-by: Paul Wicking <paul.wicking@qt.io>