qdoc: Introduce automatic generation of navigation links
Introduce a new configuration variable, navigation.toctitles, for
automatic generation of navigation links for pages listed in a TOC.

A TOC page is a \page with a \list, where list elements define the pages
included in the documentation set. When set, QDoc will generate links
for these pages, without the need for \nextpage and \previouspage
commands.

[ChangeLog][qdoc] Added a capability to generate navigation links for
pages listed in a TOC.

Fixes: QTBUG-84238
Change-Id: I0b34d8a0fe4b01a2ffcc084e6416c0a4b4623b78
Reviewed-by: Paul Wicking <paul.wicking@qt.io>