Fix MSVC runtime detection for WinRT debug binaries
Prior to the fix if the app depends on any library with 'app' substring
in the name the build will always be detected as release. For example
WinRT apps which depend on vccorlib_app library will always be detected as release.

Change-Id: Ib17376962f3d5120211338b4baa08a431a226dd7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit 096dfcf4f0d153546ca2bad4191b471c97731970)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>