CMake: Exclude a few more apps when doing iOS builds
The apps fail to build in a device_and_simulator configuration due to
trying to link a libz file by absolute path, which is missing certain
architecture slices. qmake instead links via -lz instead of an
absolute path, but we haven't figured out how to do that nicely
yet in CMake.

The apps are somehow excluded by qmake as well (allthough for a
different reason), unfortunately I could not yet pinpoint how it's
done, except for the fact that load(qt_app) does it.

Use the existing exclusion mechanism as a shortcut to get
device_and_simulator configurations in Coin ASAP.

Task-number: QTBUG-80899
Change-Id: I72724b86b3086473683e0a50e6f232fdaa94c48b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>