CMake: Use LLVM_INSTALL_DIR env and cache var when looking for clang
For convenience and compatibility with what Coin currently sets,
instead of forcing a user to set CMAKE_PREFIX_PATH to point to the
llvm dir path (and thus override any previously set path, perhaps by a
toolchain file), allow usage of either the LLVM_INSTALL_DIR
environment variable or CMake cache variable.

Make sure to use the value for CMAKE_FIND_ROOT_PATH as well, in case
if doing cross-builds, and the path is not in the sysroot.

Change-Id: I6927f14d7f721a018323225f6276a2c388dfcf20
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>