qdoc: Extend synopsis for detailed reference sections
Include \since information for detailed synopsis generation,
and allow duplication of qdoc comments for C++/QML nodes
on the condition that the extra comments contain no
body text, only meta-commands.

The use case is to have full status information in the
signature/section title, and allow additional meta-commands
for items that share a documentation comment.

/*!
    \fn void A:foo()
    \fn void A:foo(int i)

    Shared documentation for overloads.
*/

/*!
    \fn void A:foo(int i)
    \since 2.0
*/

Extract the code for 'extra' synopsis into a new helper function
and use for both HTML and DocBook generators.

Fixes: QTBUG-55739
Fixes: QTBUG-82902
Change-Id: I6e797d3ce7bc7b02102bb949515f88e7a0fbe929
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>