CMake: Fix configure.cmake usage for libclang
Now that the qmake situation of the configure.json files in regards to
libclang is fixed, we need to adjust the CMake files to do the same.

Remove the old stale configure.cmake files, create the top-level
configure.cmake file, evaluate its features in src dir, and make sure
to also create the forwarding headers in the global QtTools module as
well.

Change FindWrapLibClang to accept a version number, and modify
configure.cmake to pass the minimum version to be "8".

While libclang 6.x is enough to build Qt with qmake, for some reason
CMake needs at least 8.x. Explicitly specifying the minimum version
will prevent an old libclang library to be found in Coin, and thus
breaking integrations during compilation.
Once Coin is updated to include latest libclang, qdoc and lupdate
will be built as usual.

Change-Id: Iab8ec25d256358005850f25d6395d81e0068edfe
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>