qdoc: Prefer sorting nodes based on the name
\generatelist and \annotatedlist commands sort the elements using
Node::nodeNameLessThan(). For node types that are not page nodes
(they do not generate a page of their own) or functions, the sorting
function considered the location (i.e, file path of the declaration)
over the actual name of the node/element.

This is almost always incorrect as the preferred sort order is
alphabetical, and the location is typically not visible in the
user-visible docs.

Pick-to: 5.15
Fixes: QTBUG-85370
Change-Id: I37ae63e34fef966e3863749b3f0cafd9bc230ca0
Reviewed-by: Andy Shaw <andy.shaw@qt.io>