qtattributionsscanner: Add property CopyrightFile
JSON doesn't really support linebreaks in strings. So far we do allow
them in qt_attribution.json files though because Qt's JSON parser
does support them.

One of the fields most often affected by this is "Copyright". Allowing
users to store multiple copyright lines in a separate file is an easy
way to deal with this.

Change-Id: I9578906f0c1cac0cfca007813feb7e926e75c730
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>