qdoc: Test the C++ property macros in QDoc's generatedoutput autotest
In order to test the Q_PROPERTY macros, we need to provide include
paths to QtCore for the test. Generate a list of include files when
running qmake or CMake, stored in a file, and load that file as
extra command line arguments when running QDoc.

Skip the test if that file is not available at the time of test
execution.

Fixes: QTBUG-85565
Change-Id: Ibe7ec4a454441c15320b3a7e1ff8ebefb44c2f5e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>