qdoc: Add support for bindable properties
Add support for the BINDABLE attribute in the Q_PROPERTY macro.

The new properties are marked with 'bindable' tag, and the list
of access functions/notifier signal is replaced with a descriptive
note and a link to QProperty. Read-only properties are also
properly marked as such.

[ChangeLog][qdoc] The \property command now supports bindable C++
properties that use the new system based on QProperty.

Task-number: QTBUG-85565
Change-Id: Ie352b3ce962b6b05a022d444da0991b8a849e474
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>