Use OpenType font weights
In Designer, the only way the user can affect the weight
of the font is by checking/unchecking "bold", and this is
stored as a separate element in the XML. It is undefined
which one takes precedence if they do not match (in fact,
uic and Designer behaves differently in this case.)

To avoid this confusion as well as keep compatibility with
old files after we changed the scale of QFont::Weight to
match that of OpenType, we just ignore the weight element
in the input file.

We also stop writing the duplicate information to the file.

Task-number: QTBUG-42248
Change-Id: I08682f697a96d764c068e47bdab5d0f45cc2150b
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>