qdoc: Fix output of \footnote command
These small changes make the output of \footnote slightly more sane.
The command is not fully implemented; for HTML generator it produces
only <!-- --> comments, but the previous behavior was incorrect and
affected formatting of the rest of the document.

With WebXML output format, \footnote command together with \link (as
in the generatedoutput test) seem to differ in linefeed placement on
Windows as compared to other platforms. Remove these from the tested
set for now.

Change-Id: Id387aaeef1de1d1a859b5ec6e388394564cd0732
Reviewed-by: Paul Wicking <paul.wicking@qt.io>