qtattributionsscanner: Linkify dejacode URN's for LicenseId
So far we only allowed SPDX license ID's. Anyhow, there are some
licenses not represented in SPDX, but in the much vaster DejaCode
database. This is the case e.g. for the Bitstream Vera Font License,
which is tracked in Qt Gui and currently leads to a 404:

   https://doc.qt.io/qt-5.15/qtgui-attribution-vera-font.html

To avoid having to flag all such licenses as 'NONE', we now allow
LicenseID to contain URN's of the form urn:dje:license:, see also
https://enterprise.dejacode.com/urn/urn:dje:license/

Pick-to: 5.15
Change-Id: Ibf42f00452dbf5d28318b3217a9fe74e33f7c97c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>