windeployqt: Add optional qmake parameter
In a crosscompilation setup (for example Windows arm64) windeployqt will
pick up the wrong Qt libraries. At the moment qmake is the only way to
determine information about the Qt build even if cmake was used to build
Qt. By passing the path to the proper qmake, windeployqt will deploy
the correct libraries for the build.

Change-Id: I14992b1c27cce774440fa16abcc2abb30bec641a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>