qdoc: ClangCodeParser: Improve function argument matching
In certain cases, Clang returns a parameter type for a function argument
that differs in the header declation and source definition. This can
happen if the type has an alias in a different namespace, both
namespaces are visible to the source file, and both the type and the
alias are included in the pre-compiled header.

Resolve these cases by recording the canonical spelling of the header
type into Parameter, and comparing it to the canonical spelling from
the source file.

Remove findFunctionNodeForCursor() as redundant. This makes the fix
apply both for /*! */ comments preceding function bodies, as well as
separate \fn commands.

Fixes: QTBUG-86665
Change-Id: I47566f90adb956095537115ff8e8fcd9c0adffbe
Reviewed-by: Paul Wicking <paul.wicking@qt.io>