lupdate: Revisit how to specify the compilation database
The compilation database's file path was passed to clang's
CommonOptionsParser with the "-p" argument. This, however, specifies a
directory, not the path to the compilation database itself.

There doesn't seem to be a way to directly specify the file path of the
compilation database itself. The whole API revolves around directories.
This is also what I observe in other clang-based tools, see for example
the -compile-commands-dir option of clangd.

Remove CommonOptionsParser, and directly use CompilationDatabase to
explicitly control how to search for the compilation database, and
actually do what the help output claims.

Reflect in the help output that lupdate expects a directory path, not a
file path.

Properly forward the error message from CompilationDatabase, and bail
out with an error exit code.

Change-Id: I935da6078d020431d0f69509da69a52089e23f8a
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>