lupdate: Clarify constness in ReadSynchronizedRef
The const/const_cast dance we do in the constructor is only needed,
because m_vector isn't a const ref. It isn't const to allow for the
reset() member function. As this class represents a reference and
references are usually not supposed to change what they point
to (*pointers* do that job), it's questionable to have the reset()
method at all.

This patch removes the reset() method and makes m_vector a const ref
again, which simplifies the class.

Change-Id: I827150b756050f5bf17bb7354315035e74a091a8
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>