Adapt and run the lupdate tests for the clang-based lupdate too
We now run the lupdate autotests once with the "classic" lupdate and
once with the clang-based lupdate, if Qt was built with clangcpp
support.
Most of the source code need to be made compilable
as the new parser can only run on compilable code.
In some cases the results for the new parser are expected
to be different. In such case a dedicated directory is created.

lupdate/clang prints warning messages with native separators.
Adjust our test data to accept paths with slashes or backslashes.

Task-number: QTBUG-87090
Change-Id: I526137f7f918461939efed57b25378109a273615
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit d012c2f69cffadb8d34733156eacf08a5a77d1be)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>