lupdate/clang: Stabilize order of translator messages
The multiple threads lupdate/clang runs lead to an unpredictable order
of translation messages in the generated .ts files. This is particularly
bad for autotests that compare the generated .ts file with expected
output.

We now collect the TranslatorMessage objects in a vector, and sort them
according to the original order of input files. Then, we sequentially
call Translator::extend with the messages.

Change-Id: Ibc6df056477fe221c4cc50f6f6c394225d340dba
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>