qdoc: Fix processing of related non-members read from index
QDoc writes a node using \relates twice into the index;
once under the original (lexical) parent (typically, the
global scope), and again under the adoptive parent specified
with \relates. This is by design, as it allows searching for
the node under both scopes.

However, when the nodes are read back in, what was originally
a single node was duplicated. This caused issues e.g. when
that node is part of a group; when generating a list for a
group, a line representing the node appeared twice.

Use Aggregate::adoptChild() when reading nodes from index,
instead of constructing multiple nodes. This is similar to
what the code parser does when processing the \relates
command.

To identify an index entry appearing twice as a representation
of a single node, repurpose the "related" attribute to track
the index of each related node.

Fixes: QTBUG-87311
Change-Id: I2aa8a97b9be26ee3945b3e4ce222b0f8dc4ca8b5
Reviewed-by: Paul Wicking <paul.wicking@qt.io>